refactor: nixed compiler and moved into jacprogram, but not done
