+++ jac/jaclang/compiler/passes/main/__init__.py
@@ -1,5 +1,11 @@
 """Collection of passes for Jac IR."""
 
+from __future__ import annotations
+
+from typing import Type
+
+from jaclang.compiler.passes import Pass
+
 from .sub_node_tab_pass import SubNodeTabPass
 from .sym_tab_build_pass import SymTabBuildPass  # noqa: I100
 from .def_use_pass import DefUsePass  # noqa: I100
@@ -15,7 +21,7 @@ from .pybc_gen_pass import PyBytecodeGenPass  # noqa: I100
 from .schedules import type_checker_sched  # noqa: I100
 
 
-pass_schedule = py_code_gen  # type: ignore[has-type]
+pass_schedule: list[Type[Pass]] = py_code_gen
 
 __all__ = [
     "SubNodeTabPass",
