+++ jac/jaclang/compiler/passes/main/tests/test_decl_def_match_pass.py
@@ -1,8 +1,8 @@
 """Test pass module."""
 
 import jaclang.compiler.absyntree as ast
-from jaclang.compiler.compile import jac_file_to_pass
 from jaclang.compiler.passes.main import DeclImplMatchPass
+from jaclang.compiler.program import JacProgram
 from jaclang.utils.test import TestCase
 
 
@@ -15,7 +15,7 @@ class DeclImplMatchPassTests(TestCase):
 
     def test_parameter_count_mismatch(self) -> None:
         """Basic test for pass."""
-        state = jac_file_to_pass(
+        state = JacProgram.jac_file_to_pass(
             self.fixture_abs_path("defn_decl_mismatch.jac"), DeclImplMatchPass
         )
 
@@ -45,7 +45,9 @@ class DeclImplMatchPassTests(TestCase):
 
     def test_ability_connected_to_decl(self) -> None:
         """Basic test for pass."""
-        state = jac_file_to_pass(self.fixture_abs_path("base.jac"), DeclImplMatchPass)
+        state = JacProgram.jac_file_to_pass(
+            self.fixture_abs_path("base.jac"), DeclImplMatchPass
+        )
         self.assertFalse(state.errors_had)
         self.assertIn("(o)Test.(c)say_hi", state.ir.sym_tab.tab)
         self.assertIsNotNone(
@@ -58,7 +60,9 @@ class DeclImplMatchPassTests(TestCase):
 
     def test_ability_connected_to_decl_post(self) -> None:
         """Basic test for pass."""
-        state = jac_file_to_pass(self.fixture_abs_path("base2.jac"), DeclImplMatchPass)
+        state = JacProgram.jac_file_to_pass(
+            self.fixture_abs_path("base2.jac"), DeclImplMatchPass
+        )
         self.assertFalse(state.errors_had)
         self.assertIn("(o)Test.(c)say_hi", state.ir.sym_tab.tab)
         self.assertIsNotNone(
@@ -71,7 +75,7 @@ class DeclImplMatchPassTests(TestCase):
 
     def test_arch_ref_has_sym(self) -> None:
         """Basic test for pass."""
-        state = jac_file_to_pass(
+        state = JacProgram.jac_file_to_pass(
             self.fixture_abs_path("defs_and_uses.jac"), DeclImplMatchPass
         )
         for i in state.ir.get_all_sub_nodes(ast.ArchRef):
@@ -79,7 +83,7 @@ class DeclImplMatchPassTests(TestCase):
 
     def test_obj_hasvar_initialization(self) -> None:
         """Basic test for pass."""
-        state = jac_file_to_pass(
+        state = JacProgram.jac_file_to_pass(
             self.fixture_abs_path("uninitialized_hasvars.jac"), DeclImplMatchPass
         )
         self.assertTrue(state.errors_had)
