+++ jac/jaclang/compiler/passes/main/tests/test_def_use_pass.py
@@ -1,7 +1,7 @@
 """Test pass module."""
 
-from jaclang.compiler.compile import jac_file_to_pass
 from jaclang.compiler.passes.main import DefUsePass
+from jaclang.compiler.program import JacProgram
 from jaclang.utils.test import TestCase
 
 
@@ -14,7 +14,7 @@ class DefUsePassTests(TestCase):
 
     def test_def_uses(self) -> None:
         """Basic test for pass."""
-        state = jac_file_to_pass(
+        state = JacProgram.jac_file_to_pass(
             file_path=self.fixture_abs_path("defs_and_uses.jac"),
             target=DefUsePass,
         )
