+++ jac/jaclang/compiler/passes/main/tests/test_pyast_gen_pass.py
@@ -6,8 +6,8 @@ import sys
 import types
 
 import jaclang.compiler.absyntree as ast
-from jaclang.compiler.compile import jac_file_to_pass
 from jaclang.compiler.passes.main import PyastGenPass, SubNodeTabPass
+from jaclang.compiler.program import JacProgram
 from jaclang.utils.test import AstSyncTestMixin, TestCaseMicroSuite
 
 
@@ -35,7 +35,7 @@ class PyastGenPassTests(TestCaseMicroSuite, AstSyncTestMixin):
 
     def test_hodge_podge(self) -> None:
         """Basic test for pass."""
-        code_gen = jac_file_to_pass(
+        code_gen = JacProgram.jac_file_to_pass(
             self.examples_abs_path("micro/hodge_podge.jac"),
             target=PyastGenPass,
         )
@@ -44,7 +44,7 @@ class PyastGenPassTests(TestCaseMicroSuite, AstSyncTestMixin):
 
     def test_circle_py_ast(self) -> None:
         """Basic test for pass."""
-        code_gen = jac_file_to_pass(
+        code_gen = JacProgram.jac_file_to_pass(
             self.examples_abs_path("manual_code/circle.jac"),
             target=PyastGenPass,
         )
@@ -86,7 +86,7 @@ class PyastGenPassTests(TestCaseMicroSuite, AstSyncTestMixin):
 
     def micro_suite_test(self, filename: str) -> None:
         """Parse micro jac file."""
-        code_gen = jac_file_to_pass(
+        code_gen = JacProgram.jac_file_to_pass(
             self.fixture_abs_path(filename), target=PyastGenPass
         )
         from_jac_str = ast3.dump(code_gen.ir.gen.py_ast[0], indent=2)
@@ -130,11 +130,11 @@ class ValidateTreeParentTest(TestCaseMicroSuite):
 
     def micro_suite_test(self, filename: str) -> None:
         """Parse micro jac file."""
-        code_gen = jac_file_to_pass(
+        code_gen = JacProgram.jac_file_to_pass(
             self.fixture_abs_path(filename), target=SubNodeTabPass
         )
         self.assertTrue(self.parent_scrub(code_gen.ir))
-        code_gen = jac_file_to_pass(
+        code_gen = JacProgram.jac_file_to_pass(
             self.fixture_abs_path(filename), target=PyastGenPass
         )
         self.assertTrue(self.parent_scrub(code_gen.ir))
