+++ jac/jaclang/compiler/passes/main/tests/test_pybc_gen_pass.py
@@ -2,7 +2,7 @@
 
 import marshal
 
-from jaclang.compiler.compile import jac_file_to_pass
+from jaclang.compiler.program import JacProgram
 from jaclang.utils.test import TestCase
 
 
@@ -15,7 +15,7 @@ class PyBytecodeGenPassTests(TestCase):
 
     def test_simple_bcgen(self) -> None:
         """Basic test for pass."""
-        jac_code = jac_file_to_pass(
+        jac_code = JacProgram.jac_file_to_pass(
             file_path=self.fixture_abs_path("func.jac"),
         )
         try:
