+++ jac/jaclang/compiler/passes/main/tests/test_sub_node_pass.py
@@ -1,7 +1,7 @@
 """Test sub node pass module."""
 
-from jaclang.compiler.compile import jac_file_to_pass
 from jaclang.compiler.passes.main import SubNodeTabPass
+from jaclang.compiler.program import JacProgram
 from jaclang.utils.test import TestCase
 
 
@@ -14,7 +14,7 @@ class SubNodePassTests(TestCase):
 
     def test_sub_node_pass(self) -> None:
         """Basic test for pass."""
-        code_gen = jac_file_to_pass(
+        code_gen = JacProgram.jac_file_to_pass(
             file_path=self.examples_abs_path("manual_code/circle.jac"),
             target=SubNodeTabPass,
         )
