+++ jac/jaclang/compiler/passes/main/tests/test_sym_tab_build_pass.py
@@ -1,6 +1,6 @@
 """Test pass module."""
 
-# from jaclang.compiler.compile import jac_file_to_pass
+# from jaclang.compiler.program import JacProgram
 from jaclang.compiler.passes.main import SymTabBuildPass
 from jaclang.utils.test import AstSyncTestMixin, TestCase
 
@@ -16,7 +16,7 @@ class SymTabBuildPassTests(TestCase, AstSyncTestMixin):
 
     # def test_name_collision(self) -> None:
     #     """Basic test for pass."""
-    #     state = jac_file_to_pass(
+    #     state = JacProgram.jac_file_to_pass(
     #         self.fixture_abs_path("multi_def_err.jac"), SymTabBuildPass
     #     )
     #     self.assertGreater(len(state.warnings_had), 0)
