+++ jac/jaclang/compiler/passes/main/tests/test_type_check_pass.py
@@ -2,8 +2,8 @@
 
 from typing import List
 
-from jaclang.compiler.compile import jac_file_to_pass
 from jaclang.compiler.passes.main.schedules import py_code_gen_typed
+from jaclang.compiler.program import JacProgram
 from jaclang.utils.lang_tools import AstTool
 from jaclang.utils.test import TestCase
 
@@ -18,7 +18,7 @@ class MypyTypeCheckPassTests(TestCase):
 
     def test_type_errors(self) -> None:
         """Basic test for pass."""
-        type_checked = jac_file_to_pass(
+        type_checked = JacProgram.jac_file_to_pass(
             file_path=self.fixture_abs_path("func.jac"),
             schedule=py_code_gen_typed,
         )
@@ -36,7 +36,7 @@ class MypyTypeCheckPassTests(TestCase):
 
     def test_imported_module_typecheck(self) -> None:
         """Basic test for pass."""
-        type_checked = jac_file_to_pass(
+        type_checked = JacProgram.jac_file_to_pass(
             file_path=self.fixture_abs_path("game1.jac"),
             schedule=py_code_gen_typed,
         )
