+++ jac/jaclang/compiler/passes/main/tests/test_typeinfo_pass.py
@@ -1,8 +1,8 @@
 """Test pass module."""
 
-from jaclang.compiler.compile import jac_file_to_pass
 from jaclang.compiler.passes.main.fuse_typeinfo_pass import FuseTypeInfoPass
 from jaclang.compiler.passes.main.schedules import py_code_gen_typed
+from jaclang.compiler.program import JacProgram
 from jaclang.utils.test import TestCase
 
 
@@ -15,7 +15,7 @@ class TestFuseTypeInfo(TestCase):
 
     def test_mod_type_assign(self) -> None:
         """Test module type assignment."""
-        gen_ast = jac_file_to_pass(
+        gen_ast = JacProgram.jac_file_to_pass(
             self.fixture_abs_path("mod_type_assign.jac"),
             FuseTypeInfoPass,
             schedule=py_code_gen_typed,
