+++ jac/jaclang/tests/test_bugs.py
@@ -1,6 +1,6 @@
 """Test Jac language generally."""
 
-from jaclang.compiler.compile import jac_file_to_pass
+from jaclang.compiler.program import JacProgram
 from jaclang.utils.test import TestCase
 
 
@@ -13,7 +13,7 @@ class JacBugTests(TestCase):
 
     def test_impl_match_confusion_issue(self) -> None:
         """Basic test for symtable support for inheritance."""
-        mypass = jac_file_to_pass(
+        mypass = JacProgram.jac_file_to_pass(
             self.fixture_abs_path("impl_match_confused.jac"),
         )
         self.assertEqual(len(mypass.errors_had), 1)
