+++ jac/jaclang/tests/test_language.py
@@ -8,7 +8,6 @@ import sysconfig
 import jaclang.compiler.passes.main as passes
 from jaclang import jac_import
 from jaclang.cli import cli
-from jaclang.compiler.compile import jac_file_to_pass, jac_pass_to_pass, jac_str_to_pass
 from jaclang.compiler.passes.main.schedules import py_code_gen_typed
 from jaclang.compiler.program import JacProgram
 from jaclang.runtimelib.context import ExecutionContext
@@ -186,7 +185,7 @@ class JacLanguageTests(TestCase):
 
     def test_arith_precedence(self) -> None:
         """Basic precedence test."""
-        prog = jac_str_to_pass("with entry {print(4-5-4);}", "test.jac")
+        prog = JacProgram.jac_str_to_pass("with entry {print(4-5-4);}", "test.jac")
         captured_output = io.StringIO()
         sys.stdout = captured_output
         exec(compile(prog.ir.gen.py_ast[0], "test.py", "exec"))
@@ -485,7 +484,7 @@ class JacLanguageTests(TestCase):
 
     def test_annotation_tuple_issue(self) -> None:
         """Test conn assign on edges."""
-        mypass = jac_file_to_pass(self.fixture_abs_path("./slice_vals.jac"))
+        mypass = JacProgram.jac_file_to_pass(self.fixture_abs_path("./slice_vals.jac"))
         self.assertIn("Annotated[Str, INT, BLAH]", mypass.ir.gen.py)
         self.assertIn("tuple[int, Optional[type], Optional[tuple]]", mypass.ir.gen.py)
 
@@ -549,7 +548,7 @@ class JacLanguageTests(TestCase):
             except Exception as e:
                 return f"Error While Jac to Py AST conversion: {e}"
 
-        ir = jac_str_to_pass(
+        ir = JacProgram.jac_str_to_pass(
             jac_str=py_ast_build_pass.ir.unparse(),
             file_path=file_name[:-3] + ".jac",
             schedule=py_code_gen_typed,
@@ -615,7 +614,7 @@ class JacLanguageTests(TestCase):
             except Exception as e:
                 return f"Error While Jac to Py AST conversion: {e}"
 
-        ir = jac_str_to_pass(
+        ir = JacProgram.jac_str_to_pass(
             jac_str=py_ast_build_pass.ir.unparse(),
             file_path=file_name[:-3] + ".jac",
             schedule=py_code_gen_typed,
@@ -674,7 +673,9 @@ class JacLanguageTests(TestCase):
             except Exception as e:
                 return f"Error While Jac to Py AST conversion: {e}"
 
-        ir = jac_pass_to_pass(py_ast_build_pass, schedule=py_code_gen_typed).ir
+        ir = JacProgram.jac_pass_to_pass(
+            py_ast_build_pass, schedule=py_code_gen_typed
+        ).ir
         architype_count = sum(
             len(mod.get_all_sub_nodes(ast.Architype))
             for mod in ir.jac_prog.modules.values()
@@ -724,12 +725,14 @@ class JacLanguageTests(TestCase):
 
     def test_py_kw_as_name_disallowed(self) -> None:
         """Basic precedence test."""
-        prog = jac_str_to_pass("with entry {print.is.not.True(4-5-4);}", "test.jac")
+        prog = JacProgram.jac_str_to_pass(
+            "with entry {print.is.not.True(4-5-4);}", "test.jac"
+        )
         self.assertIn("Python keyword is used as name", str(prog.errors_had[0].msg))
 
     def test_double_format_issue(self) -> None:
         """Basic precedence test."""
-        prog = jac_str_to_pass("with entry {print(hello);}", "test.jac")
+        prog = JacProgram.jac_str_to_pass("with entry {print(hello);}", "test.jac")
         prog.ir.unparse()
         before = prog.ir.format()
         prog.ir.format()
@@ -847,7 +850,7 @@ class JacLanguageTests(TestCase):
                     )
                 )
             settings.print_py_raised_ast = True
-            ir = jac_pass_to_pass(jac_ast).ir
+            ir = JacProgram.jac_pass_to_pass(jac_ast).ir
             gen_ast = ir.pp()
             if module_path == "random":
                 self.assertIn("ModulePath - statistics -", gen_ast)
@@ -859,7 +862,9 @@ class JacLanguageTests(TestCase):
         file_name = os.path.join(self.fixture_abs_path("./"), "random_check.jac")
         from jaclang.compiler.passes.main.schedules import py_code_gen, PyImportPass
 
-        imp = jac_file_to_pass(file_name, schedule=py_code_gen, target=PyImportPass)
+        imp = JacProgram.jac_file_to_pass(
+            file_name, schedule=py_code_gen, target=PyImportPass
+        )
         self.assertEqual(len(imp.import_table), 5)
 
     def test_access_modifier(self) -> None:
@@ -898,7 +903,9 @@ class JacLanguageTests(TestCase):
                 raise Exception(f"Error While Jac to Py AST conversion: {e}")
 
         settings.print_py_raised_ast = True
-        ir = jac_pass_to_pass(py_ast_build_pass, schedule=py_code_gen_typed).ir
+        ir = JacProgram.jac_pass_to_pass(
+            py_ast_build_pass, schedule=py_code_gen_typed
+        ).ir
         jac_ast = ir.pp()
         self.assertIn(" |   +-- String - 'Loop completed normally{}'", jac_ast)
         self.assertEqual(len(ir.get_all_sub_nodes(ast.SubNodeList)), 586)
@@ -920,7 +927,7 @@ class JacLanguageTests(TestCase):
 
     def test_ds_type_check_pass(self) -> None:
         """Test conn assign on edges."""
-        mypass = jac_file_to_pass(
+        mypass = JacProgram.jac_file_to_pass(
             self.examples_abs_path("micro/simple_walk.jac"),
             schedule=py_code_gen_typed,
         )
@@ -930,7 +937,7 @@ class JacLanguageTests(TestCase):
 
     def test_ds_type_check_pass2(self) -> None:
         """Test conn assign on edges."""
-        mypass = jac_file_to_pass(
+        mypass = JacProgram.jac_file_to_pass(
             self.examples_abs_path("guess_game/guess_game5.jac"),
             schedule=py_code_gen_typed,
         )
@@ -940,7 +947,7 @@ class JacLanguageTests(TestCase):
 
     def test_circle_override1_type_check_pass(self) -> None:
         """Test conn assign on edges."""
-        mypass = jac_file_to_pass(
+        mypass = JacProgram.jac_file_to_pass(
             self.examples_abs_path("manual_code/circle.jac"),
             schedule=py_code_gen_typed,
         )
@@ -968,7 +975,7 @@ class JacLanguageTests(TestCase):
 
     def test_multiline_single_tok(self) -> None:
         """Test conn assign on edges."""
-        mypass = jac_file_to_pass(self.fixture_abs_path("byllmissue.jac"))
+        mypass = JacProgram.jac_file_to_pass(self.fixture_abs_path("byllmissue.jac"))
         self.assertIn("2:5 - 4:8", mypass.ir.pp())
 
     @pytest.mark.xfail(
@@ -976,14 +983,14 @@ class JacLanguageTests(TestCase):
     )
     def test_single_impl_annex(self) -> None:
         """Basic test for pass."""
-        mypass = jac_file_to_pass(
+        mypass = JacProgram.jac_file_to_pass(
             self.examples_abs_path("manual_code/circle_pure.jac"),
             target=passes.JacImportPass,
         )
 
         self.assertEqual(mypass.ir.pp().count("AbilityDef - (o)Circle.(c)area"), 1)
         self.assertIsNone(mypass.ir._sym_tab)
-        mypass = jac_file_to_pass(
+        mypass = JacProgram.jac_file_to_pass(
             self.examples_abs_path("manual_code/circle_pure.jac"),
             target=passes.SymTabBuildPass,
         )
@@ -994,7 +1001,7 @@ class JacLanguageTests(TestCase):
 
     def test_inherit_baseclass_sym(self) -> None:
         """Basic test for symtable support for inheritance."""
-        mypass = jac_file_to_pass(
+        mypass = JacProgram.jac_file_to_pass(
             self.examples_abs_path("guess_game/guess_game4.jac"),
             target=passes.DefUsePass,
         )
