+++ jac/jaclang/utils/lang_tools.py
@@ -7,10 +7,10 @@ import sys
 from typing import List, Optional, Type
 
 import jaclang.compiler.absyntree as ast
-from jaclang.compiler.compile import jac_file_to_pass, jac_str_to_pass
 from jaclang.compiler.passes.main.pyast_load_pass import PyastBuildPass
 from jaclang.compiler.passes.main.schedules import py_code_gen, type_checker_sched
 from jaclang.compiler.passes.main.schedules import py_code_gen_typed
+from jaclang.compiler.program import JacProgram
 from jaclang.compiler.symtable import SymbolTable
 from jaclang.utils.helpers import auto_generate_refs, pascal_to_snake
 
@@ -227,7 +227,7 @@ class AstTool:
                         ),
                     ).ir
 
-                    ir = jac_str_to_pass(
+                    ir = JacProgram.jac_str_to_pass(
                         jac_str=rep.unparse(),
                         file_path=file_name[:-3] + ".jac",
                         schedule=py_code_gen_typed,
@@ -235,7 +235,7 @@ class AstTool:
                 except Exception as e:
                     return f"Error While Jac to Py AST conversion: {e}"
             else:
-                ir = jac_file_to_pass(
+                ir = JacProgram.jac_file_to_pass(
                     file_name, schedule=[*(py_code_gen), *type_checker_sched]
                 ).ir
